langgraph/roast-agent/requirements.txt
python-dotenv
fastapi
uvicorn
langchain
langchain-community>=0.0.13
langchain-openai
langgraph
pydantic>=2.0.0
beautifulsoup4
httpx
streamlit
duckduckgo-search